+++ packages/runtime-core/src/apiCreateApp.ts
@@ -241,7 +241,7 @@ export function createAppAPI<HostElement>(
       provide(key, value) {
         if (__DEV__ && key in context.provides) {
           warn(
-            `App already provides property with key "${key}". ` +
+            `App already provides property with key "${String(key)}". ` +
               `It will be overwritten with the new value.`
           )
         }