+++ packages/vue/src/devCheck.ts
@@ -1,5 +1,5 @@
 if (__BROWSER__ && __DEV__) {
-  console[console.info ? 'info' : 'log'](
+  console[(console as any).info ? 'info' : 'log'](
     `You are running a development build of Vue.\n` +
       `Make sure to use the production build (*.prod.js) when deploying for production.`
   )